renamed stop_fsm.h, stop_fsm.cpp, stop_fsm_test.cpp and the stop directory to use halt instead of stop
