+++ src/software/ai/hl/stp/play/halt_play.cpp
@@ -1,6 +1,6 @@
 #include "software/ai/hl/stp/play/halt_play.h"
 
-#include "software/ai/hl/stp/tactic/stop/halt_tactic.h"
+#include "software/ai/hl/stp/tactic/halt/halt_tactic.h"
 #include "software/util/generic_factory/generic_factory.h"
 
 HaltPlay::HaltPlay(TbotsProto::AiConfig config) : Play(config, false) {}
