+++ src/software/ai/hl/stp/tactic/halt/halt_fsm.cpp
@@ -1,0 +1,11 @@
+#include "software/ai/hl/stp/tactic/halt/halt_fsm.h"
+
+void HaltFSM::updateHalt(const Update& event)
+{
+    event.common.set_primitive(createStopPrimitive());
+}
+
+bool HaltFSM::haltDone(const Update& event)
+{
+    return robotStopped(event.common.robot);
+}
