+++ src/software/ai/hl/stp/tactic/halt/halt_tactic.cpp
@@ -1,4 +1,4 @@
-#include "software/ai/hl/stp/tactic/stop/halt_tactic.h"
+#include "software/ai/hl/stp/tactic/halt/halt_tactic.h"
 
 #include <algorithm>
 
@@ -6,7 +6,7 @@ HaltTactic::HaltTactic() : Tactic(std::set<RobotCapability>()), fsm_map()
 {
     for (RobotId id = 0; id < MAX_ROBOT_IDS; id++)
     {
-        fsm_map[id] = std::make_unique<FSM<StopFSM>>(StopFSM());
+        fsm_map[id] = std::make_unique<FSM<HaltFSM>>(HaltFSM());
     }
 }
 
@@ -19,8 +19,8 @@ void HaltTactic::updatePrimitive(const TacticUpdate &tactic_update, bool reset_f
 {
     if (reset_fsm)
     {
-        fsm_map[tactic_update.robot.id()] = std::make_unique<FSM<StopFSM>>(StopFSM());
+        fsm_map[tactic_update.robot.id()] = std::make_unique<FSM<HaltFSM>>(HaltFSM());
     }
     fsm_map.at(tactic_update.robot.id())
-        ->process_event(StopFSM::Update({}, tactic_update));
+        ->process_event(HaltFSM::Update({}, tactic_update));
 }
