--- src/software/ai/hl/stp/tactic/stop/stop_fsm.cpp
@@ -1,11 +1,0 @@
-#include "software/ai/hl/stp/tactic/stop/stop_fsm.h"
-
-void StopFSM::updateStop(const Update& event)
-{
-    event.common.set_primitive(createStopPrimitive());
-}
-
-bool StopFSM::stopDone(const Update& event)
-{
-    return robotStopped(event.common.robot);
-}
